requirements.txt
modbus-tk
pyserial
Flask